requirements.txt
Django==1.9.4
psycopg2==2.6.1
pytest-django==2.9.1
python-dateutil==2.5.1
sure==1.2.24
django-herokuapp==0.9.20
